+++ src/controllers/LicensesController.php
@@ -28,9 +28,6 @@ class LicensesController extends BaseController
         parent::init();
     }
 
-    // Public Methods
-    // =========================================================================
-
     /**
      * Create or edit a License
      *