+++ src/migrations/Install.php
@@ -15,9 +15,6 @@ use craft\helpers\MigrationHelper;
  */
 class Install extends Migration
 {
-    // Public Methods
-    // =========================================================================
-
     /**
      * @inheritdoc
      */
@@ -42,9 +39,6 @@ class Install extends Migration
 
         return true;
     }
-
-    // Protected Methods
-    // =========================================================================
 
     /**
      * Creates the tables for Digital Products